# static/well-known/openapi.yaml
openapi: 3.1.0
info:
  title: LifeOS Co-Pilot API
  version: 1.0.0
  description: API for managing memory, tasks, and general queries for LifeOS Co-Pilot.

servers:
  - url: https://life-os-private-practical-co-pilot.onrender.com

paths:
  /ask:
    post:
      operationId: askQuestion
      summary: Send a general question to the LifeOS Co-Pilot.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  default: austin_master
                prompt:
                  type: string
              required:
                - prompt
      responses:
        '200':
          description: The AI response
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string

  /memory:
    get:
      operationId: getMemory
      summary: Retrieve stored memory
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            default: austin_master
      responses:
        '200':
          description: Memory retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  memory:
                    type: string

    post:
      operationId: addMemory
      summary: Store memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  default: austin_master
                memory:
                  type: string
              required:
                - memory
      responses:
        '201':
          description: Memory stored

  /tasks:
    get:
      operationId: getTasks
      summary: List tasks
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            default: austin_master
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

    post:
      operationId: createTask
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  default: austin_master
                task:
                  type: string
              required:
                - task
      responses:
        '201':
          description: Task created

  /tasks/{id}:
    patch:
      operationId: updateTask
      summary: Update an existing task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  default: austin_master
                task:
                  type: string
              required:
                - task
      responses:
        '200':
          description: Task updated

    delete:
      operationId: deleteTask
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  default: austin_master
      responses:
        '204':
          description: Task deleted successfully

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key